pyproject.toml
[project]
name = "meshy-research-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "dashscope>=1.25.13",
    "edge-tts>=7.2.7",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
